all wired up to main, but missing firebase app, I'll implement as it's own feature
