+++ integrations/example/lib/example_feature.dart
@@ -1,3 +1,4 @@
+import 'package:dart_board_authentication_flutter_fire/dart_board_authentication_flutter_fire.dart';
 import 'package:dart_board_core/dart_board.dart';
 import 'package:dart_board_core/impl/features/generic_features.dart';
 import 'package:dart_board_minesweeper/dart_board_minesweeper.dart';
@@ -50,6 +51,7 @@ class ExampleFeature extends DartBoardFeature {
         MinesweeperFeature(),
         FireCursorFeature(),
         RainbowCursorFeature(),
+        DartBoardAuthenticationFlutterFireFeature(),
 
         /// Add 2 template's
         /// can toggle in debug
