+++ integrations/example/lib/examples/authentication.dart
@@ -1,0 +1,4 @@
+import 'package:dart_board_authentication/example.dart' as auth_main;
+
+/// We redirect to the Auth Example here
+void main() => auth_main.main();
