--- integrations/example/lib/examples/authentication_example.dart
@@ -1,3 +1,0 @@
-import 'package:dart_board_authentication/example.dart' as AuthMain;
-
-void main() => AuthMain.main();
