+++ integrations/example/pubspec.yaml
@@ -36,6 +36,8 @@ dependencies:
   dart_board_particles:
   dart_board_authentication: 
     path: ../../features/dart_board_authentication
+  dart_board_authentication_flutter_fire: 
+    path: ../../features/dart_board_authentication_flutter_fire
   
   flutter_syntax_view: ^4.0.0
   flutter_markdown: ^0.6.2
